Fix array dtype

+++ artistools/spectra.py
@@ -46,7 +46,7 @@ fluxcontributiontuple = namedtuple(
 def stackspectra(spectra_and_factors):
     factor_sum = sum([factor for _, factor in spectra_and_factors])
 
-    stackedspectrum = np.zeros_like(spectra_and_factors[0][0])
+    stackedspectrum = np.zeros_like(spectra_and_factors[0][0], dtype=np.float)
     for spectrum, factor in spectra_and_factors:
         stackedspectrum += spectrum * factor / factor_sum
